Halted attempt at MAAS2 maasObjectNetworkInterfaces
My first thought was to share the big loop that constructs the
network.InterfaceInfos and make the maasInterface constellation of
types conform to the corresponding gomaasapi interfaces.

I've decided that squashes them together too much - if we can't mash the
api1 types into the api2 interfaces then we'll still need to branch. And
it was leaking out into the other places where the local maasInterface
types are used. Eventually we'll need to fix those bits too, but this
way we don't need to do all of them at once.