Add RemoteApplication and -Endpoint to description
For now these will be serializable but not importable - we don't want to
accidentally allow a migration to succeed while not migrating remote
applications from the source model.

Handle remote applications in model.Validate. Since there are no units
in the local model for a remote application, we can't check that the
settings for the endpoints and units match up.