worker/uniter: Remove relation state dir even if non-empty
As part of the fix for bug #1699050, we added an upgrade step to remove
spurious relation scopes. In some cases (which I haven't managed to
reproduce unfortunately), the uniter state directory has a file under
the relation directory for a spurious relation. As the uniter restarts
it cleans up relation directories that are no longer relevant, but fails
because the directory isn't empty. Use os.RemoveAll to always remove the
directory (and log if it wasn't empty).