Abortive attempt at converting auditlog test to featuretest
Tried to move MachineSuite.TestManageModelServesAPI to a feature test
since the behaviour is in the worker/apiserver in 2.4, but couldn't get
it working. Current status is that it looks like StateServingInfo isn't
set in the agent config so the API server never gets started in the
agent. Bailing for now.