audit-log: always drop excluded methods
It's confusing to explain the behaviour of audit log filtering when it
includes all messages in an interesting conversation (even ones which
are in the exclude list), which means it's probably wrong. Change it so
that it always excludes those messages.

Tracks the IDs of requests it's dropped so it can also drop the
responses.