Add benchmark-leadership command
This will report statistics about how many leadership claim requests a
controller can support per second. From running it against a controller
in aws (in ap-southeast-2) I see about 23 requests/sec for 10 units, but
this falls down to about 16 per second for 35 units (or if we increase
the number of goroutines for a specific unit).