Reworked benchmark to be a standalone binary
This means it can be run on a machine in the model (where the juju
client isn't installed).

Extended the config format to include the values we were getting from
the juju client store. And allow for running the leadership claim
against only some of the units in the config.