Experiment: ensure connection goes to raft leader
Change api.Info to allow users to say they only want a connection to the
raft leader, and change the API to handle that. This would enable the
client to be able to use basically the same connection logic without
having to complicate the controller.

Unfortunately it looks like this would require either keeping 2 api
connections open per agent (with half of them being to the raft leader),
or opening a connection each time we wanted to make a leadership
claim. The former is probably a non-starter, and from some benchmarking
the latter seems like it would swamp the speedup that moving to raft
would give. It's possible that this is mostly client-side cost? But
pausing this approach for now.