Fix goroutine and websocket leak in logsink
If the ServeHTTP handler finishes before the receiveLogs before the
receiveLogs goroutine, it can be stuck in the select trying to send on
logCh. Add a stop channel to the goroutine that is always closed when
the handler finishes and include that in the select.

This should fix the leak we see in prodstack, although we don't
understand why it's happening now - none of the logsink code has changed
between 2.4 and 2.5