checkpoint
This is a change to increase the performance of the lease manager when
there are lots of clients talking to it by reducing the number of times
it gets all of the leases at once. This requires copying a large map
with a lock held, and can cause all claims/extends to queue. To avoid
that we only get all leases once a second for nextTick and checking
blocks. Other changes to only get specific lease info and avoid doing
expiry checking for stores that autoexpire should also help.

This just needs tests, but I'm parking it for now - @jameinel had an
idea for a bigger change that might have much better improvement.